feat: refactor component imports and add AboutSection
