+++ packages/landing/src/components/InteractiveBox/index.tsx
