+++ packages/landing/src/components/Logo/index.tsx
