+++ packages/landing/src/global/spacing.scss
@@ -7,3 +7,4 @@ $spacing-large: 24px;
 $spacing-big: 32px;
 $spacing-huge: 40px;
 $spacing-massive: 48px;
+$spacing-colossal: 64px;
